dev summarize.sh

+++ c/ou/python/parse_output_remote.py
@@ -84,6 +84,27 @@ def parse_log(input_fn,qfoo,include_inputs=True,printing=False):
     # df.drop(columns=['dt','reflect','set_second'],inplace=True)
     return df
 
+# ##################################
+# # (General) Parse Runtime
+# ##################################
+# import datetime, os, time
+# import dateutil.parser as dp
+# def parse_runtime(input_fn):
+#     seconds_in_day = 24 * 60 * 60
+#     seconds_in_year = seconds_in_day * 365
+#     with open(input_fn) as f:
+#         line_first = f.readline()
+#         for n, line in enumerate(f):
+#             line_last = f.readline()
+#     # import dateutil.tz as tz
+#     time_start = dp.parse(line_first[:-1], ignoretz=True)
+#     time_end = dp.parse(line_last[:-1], ignoretz=True)
+#     difference = time_end - time_start
+#     # divmod(difference.days * seconds_in_day + difference.seconds, 60)#(min,sec)
+#     seconds = difference.days * seconds_in_day + difference.seconds
+#     years = seconds / seconds_in_year
+#     return years
+
 ##################################
 # For Dense Results
 ##################################
@@ -136,11 +157,14 @@ def parse_output_log(input_fn,include_inputs=True, printing=False):
     return df
 
 if __name__=='__main__':
-    
+    import sys
+    input_fn=sys.argv[1]
+    sum_fn=sys.argv[2]
     # input_folder=f"../data"
     # os.chdir(input_folder)
-    input_fn='../Log/example-output.txt'
+    # input_fn='../Log/example-output.txt'
     assert (os.path.exists(input_fn))
+    assert (os.path.exists(sum_fn))
     n_input,n_output=parse_iolines(input_fn, printing=False)
     df=parse_outputs(input_fn, n_output)
     df.set_index('N',inplace=True,drop=True)
@@ -148,16 +172,33 @@ if __name__=='__main__':
     # print(f"is this a pandas.Dataframe? {type(df)}") Yes
     print(df.head())
 
-    #TODO: argument parse for seed   
-    seed  =1234
-    sum_fn="Log/"+seed+".csv"
+    #TODO(later): argument parse for seed (the original)
+    # seed  =1234
+    # sum_fn=f"sum.csv"
+    boo=df.CollTime!=-9999
+    if not os.path.exists(sum_fn):
+        #save initial sum/summary
+        count=0
+        df['count']=0
+        if count==0:
+            df.loc[boo,'CollTime']=df[boo]['CollTime']+df[boo]['CollTime']
+            df.loc[boo,'count']+=1
+            df.to_csv(sum_fn)
+        # print(df.head())
+    else:
+        #appendd initial sum/summary:
+        df_sum=pd.read_csv(sum_fn,index_col="N")
+        boo_sum=df_sum.CollTime!=-9999
+        df_sum.loc[boo_sum,'CollTime']=df_sum[boo_sum]['CollTime']+df[boo]['CollTime']
+        df_sum.loc[boo_sum,'count']=1+df_sum.loc[boo_sum,'count']
+        df_sum.to_csv(sum_fn)
 
-    #save initial sum/summary
-    count=0
-    if count==0:
-        df.to_csv(sum_fn)
 
-    print(df.head())
+    #reconstruct my preexisting input
+    #TODO: read input_fn
+    #TODO: print each line in input_fn up until n_output
+    #TODO: load and sum an input file with sum_fn,
+    #TODO: save df to sum_fn
 
     # # Test compute_runtime_for_folder
     # nb_dir='/home/timothytyree/Documents/GitHub/bgmc/python/'
